requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
requests>=2.28.0
google-genai>=0.1.0
pytz>=2023.3
